:hammer: relocate the python356 exception to add_a_new_function

+++ test/test_javascript.py
@@ -18,27 +18,36 @@ def label_formatter(params):
     return params.name + 'abc'
 
 
-def test_label_formatter():
+def generic_formatter_t_est(**keywords):
     attr = ["Jan", "Feb"]
     v1 = [2.0, 4.9]
     bar = Bar("Bar chart", "precipitation and evaporation one year")
-    bar.add("precipitation", attr, v1, mark_line=["average"],
-            mark_point=["max", "min"], label_formatter=label_formatter)
     if PY35_ABOVE:
         if WINDOWS:
             with assert_raises(exceptions.ExtensionMissing):
-                bar.render()
+                bar.add("precipitation", attr, v1, mark_line=["average"],
+                        mark_point=["max", "min"],
+                        **keywords)
         else:
+            bar.add("precipitation", attr, v1, mark_line=["average"],
+                    mark_point=["max", "min"], **keywords)
             bar.render()
-            content = get_default_rendering_file_content()
-            assert 'function label_formatter(params)' in content
-            assert 'params.name + \"abc\"' in content
-            assert '"formatter": label_formatter' in content
-            os.unlink('render.html')
     else:
         with assert_raises(exceptions.JavascriptNotSupported):
-            bar.render()
+            bar.add("precipitation", attr, v1, mark_line=["average"],
+                    mark_point=["max", "min"], **keywords)
+
     javascript.clear()
+
+
+def test_label_formatter():
+    generic_formatter_t_est(label_formatter=label_formatter)
+    if PY35_ABOVE and not WINDOWS:
+        content = get_default_rendering_file_content()
+        assert 'function label_formatter(params)' in content
+        assert 'params.name + \"abc\"' in content
+        assert '"formatter": label_formatter' in content
+        os.unlink('render.html')
 
 
 def yaxis_formatter(value, index):
@@ -46,26 +55,13 @@ def yaxis_formatter(value, index):
 
 
 def test_yaxis_formatter():
-    attr = ["Jan", "Feb"]
-    v1 = [2.0, 4.9]
-    bar = Bar("Bar chart", "precipitation and evaporation one year")
-    bar.add("precipitation", attr, v1, mark_line=["average"],
-            mark_point=["max", "min"], yaxis_formatter=yaxis_formatter)
-    if PY35_ABOVE:
-        if WINDOWS:
-            with assert_raises(exceptions.ExtensionMissing):
-                bar.render()
-        else:
-            bar.render()
-            content = get_default_rendering_file_content()
-            assert 'function yaxis_formatter(value, index)' in content
-            assert 'value + index' in content
-            assert '"formatter": yaxis_formatter' in content
-            os.unlink('render.html')
-    else:
-        with assert_raises(exceptions.JavascriptNotSupported):
-            bar.render()
-    javascript.clear()
+    generic_formatter_t_est(yaxis_formatter=yaxis_formatter)
+    if PY35_ABOVE and not WINDOWS:
+        content = get_default_rendering_file_content()
+        assert 'function yaxis_formatter(value, index)' in content
+        assert 'value + index' in content
+        assert '"formatter": yaxis_formatter' in content
+        os.unlink('render.html')
 
 
 def xaxis_formatter(value, index):
@@ -73,26 +69,13 @@ def xaxis_formatter(value, index):
 
 
 def test_xaxis_formatter():
-    attr = ["Jan", "Feb"]
-    v1 = [2.0, 4.9]
-    bar = Bar("Bar chart", "precipitation and evaporation one year")
-    bar.add("precipitation", attr, v1, mark_line=["average"],
-            mark_point=["max", "min"], xaxis_formatter=xaxis_formatter)
-    if PY35_ABOVE:
-        if WINDOWS:
-            with assert_raises(exceptions.ExtensionMissing):
-                bar.render()
-        else:
-            bar.render()
-            content = get_default_rendering_file_content()
-            assert 'function xaxis_formatter(value, index)' in content
-            assert 'value + index' in content
-            assert '"formatter": xaxis_formatter' in content
-            os.unlink('render.html')
-    else:
-        with assert_raises(exceptions.JavascriptNotSupported):
-            bar.render()
-    javascript.clear()
+    generic_formatter_t_est(xaxis_formatter=xaxis_formatter)
+    if PY35_ABOVE and not WINDOWS:
+        content = get_default_rendering_file_content()
+        assert 'function xaxis_formatter(value, index)' in content
+        assert 'value + index' in content
+        assert '"formatter": xaxis_formatter' in content
+        os.unlink('render.html')
 
 
 def tooltip_formatter(params):
@@ -100,23 +83,10 @@ def tooltip_formatter(params):
 
 
 def test_tooltip_formatter():
-    attr = ["Jan", "Feb"]
-    v1 = [2.0, 4.9]
-    bar = Bar("Bar chart", "precipitation and evaporation one year")
-    bar.add("precipitation", attr, v1, mark_line=["average"],
-            mark_point=["max", "min"], tooltip_formatter=tooltip_formatter)
-    if PY35_ABOVE:
-        if WINDOWS:
-            with assert_raises(exceptions.ExtensionMissing):
-                bar.render()
-        else:
-            bar.render()
-            content = get_default_rendering_file_content()
-            assert 'function tooltip_formatter(params)' in content
-            assert 'params.name + \"abc\"' in content
-            assert '"formatter": tooltip_formatter' in content
-            os.unlink('render.html')
-    else:
-        with assert_raises(exceptions.JavascriptNotSupported):
-            bar.render()
-    javascript.clear()
+    generic_formatter_t_est(tooltip_formatter=tooltip_formatter)
+    if PY35_ABOVE and not WINDOWS:
+        content = get_default_rendering_file_content()
+        assert 'function tooltip_formatter(params)' in content
+        assert 'params.name + \"abc\"' in content
+        assert '"formatter": tooltip_formatter' in content
+        os.unlink('render.html')
